Update Leadership and 28 Secondary Skills
Warning: This update includes the two options with ini settings: Leadership and 28 Secondary Skills. Merging the changes is expected.

Leadership:
Now leadership uses json as its localization file

28 Secondary Skills:
Now the script is fully compatible with "no secondary skill limit.bin" of ERA. When the bin patch is enabled, the script will disabled itself, as well as when 10SS is enabled.