Multiple fixes
Warning: This update requires the localization of both WS Rus and ES Rus to be updated, including wog scripts.pac - zsetup01.txt and script72.ers.

Details of the localization:
WS: Alt-click to resort hero and town list instead of Shift-click
Shift-clicking on an owned external dwelling in order to transfer owner instead of using chat bar

ES: Shift-clicking on an owned dwelling to give away ownership, instead of clicking on the entrnace.

Other non-localization related changes:
- Adds back missing files from E s E
- Now Refusal of dwelling will disable the dwelling feature of transfer owner