WS imporvements
- Fixed Cards of Prophecy not giving additional mana when Knowledge is boost by the artifact
- Reduced the costs needed for Gold Reserve and Dwelling upgrades.
Now The upgrade of dwelling is calculated by the non-upgraded creature of the dwelling, this means with dirty hacks in TUM, the costs will remained the same before and after building TUM building (previously the costs would be changed due the UN:T changes). The costs of dwelling upgrade is also reduced with this new calculation.
Gold Reserve no longer requires all 4 rare resources. Now it requires Wood + Ore + Gold + a special resource. Special resource usually is Mithril. If Mithril is not enabled, the special resource would be replaced with the special resource cost of the level 7 units of the town.