Multiple fixes
WS:
- Fixed Enhanced Artifact I incorrect use of v1, which resulted in functions not executed.
Note: The use of v1 should be avoided whenever there is a chance as v1 is easily override by internal function executions - advised by Berserker.
- Remove related stuff of Display Hidden Skills as the option has been removed.

ES:
- Now you are not able to occupy your ally's mills/gardens, just like the original H3 mechanism.
Note: The script is still buggy as the original mill/garden visiting behaviours have not yet to be disabled. @daemon1995 should fix this.