Multiple improvements
ES:
Grand Manouvre: Fixed a type which results in the script not working at all.
Use of Scrolls and Banners: Fixed the calculation of exp from sacrificing artifacts

GEM:
Added a new script for "fixing" the damaging target of Magog attacks. Now Magog's Fireball respects the Fire Immunity flag of indirect targets.

ACM:
Added a check for valid battle stack in Switching action as a player reported there was a problem.