Fixes and rebalance
WS:
- Fixed the logic of "Fixing Warfare" level. Now the function look for the skill that has different level with the others to set as the new Warfare level.
- Enhanced Commander Artifacts:
Now you will not lose your Expert SS when taking off your commander artifact that gives a level to that SS (same implementation as the original ws, it would need to use much more vars the other way).
Enhanced commander Arts now give 10 different skills to the 10 arts. If Warfare is enabled, none of the war machines skills would be chosen as the commander art SS (for balance and issueless concern).
- Looks like Warfare is now fixed?

ES:
Correct the credit of JackSlater's unrestricted Morale and Luck