Multiple fixes
WS:
- First Aid I: Now the message of resurrecting dead troops after battle will only be displayed to the winner (if human).
- Eagle Eye I: Optimised the dialogue for learning a spell when there is only one spell to learn
- Mysticism I: Fixed not regenerating % of max mana after getting more than 127 power.
- Enhanced War Machine I:
a. Rewrote the whole Catapult feature with method from WM II. The fireball of Catapult is now cast after attack. The damage of the Fireball is based on the hero's Ballistics level.
b. Now you are able to upgrade multiple times of a War Machine in a town.
c. Catapult will appear in dwelling battles if the hero learns Ballistics.
d. Possibly fixed the issue the names of WM messing up.
- Living Scrolls now works only in battles vs Neutrals Creatures (as it was before).

ES and TrainerX:
- code optimisation

Problem yet to be solved:
- Make Catapult shoot even when enemy adjacent (like Ballista)