Update new game difficulty.erm
Fixed the way of fixing overflow.
Note that when the value overflow, it can become negative, and thus VR:F will only set the value to the minimum value, while we want the max. VR:S fixes it.
The max value is 1500000000 - slightly smaller than IN_MAX, this is because there might be other bonuses for the commander, leaving a bit of space is safer.