Added new creature description feature for Emerald Tower
Emerald Tower:
New creature description would be added if a new ability is added to the creature by Emerald Tower.

Note that it takes up to 5 lines of creature descriptions for Emerald Tower enhancement.
There are in total of 9/10 lines. (10 is not ideal as the last line is partlally out of the field)
This means only 4 lines of creature description is allowed before the Emreald Tower Enhancement.
If this can be a problem, I will cut out a new line character here.
Normal stats are not added to display as they require complicated string management, which isn't erm capable of.