Update 14 wog - tobyn's scripts.erm
Disable debug msg on HE:S execution to disable erm unknown error (bug report: https://discord.com/channels/665742159307341827/1110648416427577406).

Note that this is only a temp solution! The real fix should be implemented later.
It is helpful to see whether the HE:S unknown error is connected to Warfare. If we still encounter the same error, then the source of evil might be something else.