WS fixes
- Enhanced Commanders:
Optimise the behaviour of Opal - now it will try to teach the commander a skill not yet learned, instead of pure random.
Fixed the tooltip when visiting Rubies, Sapphires and Emeralds when the commander cannot learn any more skill.
Fixed
Optimise the dialogue of learning a new skill, now with pictures of the skills.
- Week of Monsters: Added 3rd mod compatibility lines