Improve mithril replacement implementation
Note:
In old wog scripts, there is an random interval between two resource pile replacement. This helps mithril to be placed on the map to be "less random". It's not likely to see a bunch of mithril stuck in one place.
In daemon's implementation, every pile of resource has VR:R chance to be replaced. This increase a lot in randomness of mithril replacement event and thus it's more likely the result is not satisfying

This commit restores the old wog scripts' random roll while still work on !?PI
The chance of replacing resource pile and campfire is slightly lower than what it was in old wog scripts