Multiple fixes
Important: It's found that in some cases in the battle with random heroes (hero aiding the neutral stacks), flag 1000 can be FALSE despite human player participate the battle. It's recommended to use i^battle_human_0^ to check whether the attacker is human to avoid the issue of flag 1000.
The issue of flag 1000 must be sorted out with Berserker later.

WS:
- Karmic Battles: Use i^battle_human_0^ instead of flag 1000 to make sure script works correctly.
- some text adaption to json

Diff Mod:
- Custom spell buffs: Use i^battle_human_0^ instead of flag 1000 to make sure script works correctly.

ES:
- Skeleton with artifacts: Improved the description of the guardian Skeleton Warriors and Skeletons on the battlefield. Now you can read the No Retaliation/Strike Twice/Strike all around they obtained.
Note that all the temporary string used in battle from Era Scritps Mod must be added to i^es_battleStrings^ and get the pointer to the string via FU(ES_AllocBattleStr). Check out the details in 1000 es - stdlib.erm